refactor:move proxy exceptions to package infra
