+++ app/adapters/repositories/strategy/proxies/proxy_method.py
@@ -2,7 +2,7 @@ from typing import List, Callable
 from app.config.messages import ERR_SECURE_PROXY as ERR_MSG
 from app.infraestructure.logging import logger
 from app.utils.decorators import cached
-from .exceptions import RestrictedAccessError
+from app.infraestructure.exceptions.proxy import RestrictedAccessError
 
 
 class MethodAccessibleProxy:
