+++ app/infraestructure/exceptions/proxy.py
